Move streamfield block definitions into separate module
